refactor: apply babel
