+++ package.json
@@ -4,19 +4,18 @@
   "private": true,
   "scripts": {
     "start": "node ./bin/www",
-    "build": "babel src/dashboard/components/src --out-dir src/dashboard/components --presets react-app/prod && webpack --env.NODE_ENV=production",
-    "build:dev": "babel src/dashboard/components/src --out-dir src/dashboard/components --presets react-app/prod && webpack --env.NODE_ENV=development"
+    "build": "webpack --env.NODE_ENV=production",
+    "build:dev": "webpack --env.NODE_ENV=development"
   },
   "dependencies": {
     "@material-ui/core": "^4.10.1",
     "@material-ui/icons": "^4.9.1",
     "@shopify/draggable": "^1.0.0-beta.8",
     "axios": "^0.19.0",
-    "babel-cli": "^6.26.0",
-    "babel-preset-react-app": "^3.1.2",
     "classnames": "^2.2.6",
     "connect-redis": "^3.4.1",
     "cookie-parser": "~1.4.3",
+    "core-js": "^3.6.5",
     "debug": "~2.6.9",
     "express": "~4.16.0",
     "express-session": "^1.16.1",
@@ -30,10 +29,15 @@
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "redis": "^2.8.0",
+    "regenerator-runtime": "^0.13.5",
     "whatwg-fetch": "^3.0.0"
   },
   "devDependencies": {
+    "@babel/core": "^7.10.2",
+    "@babel/preset-env": "^7.10.2",
+    "@babel/preset-react": "^7.10.1",
     "@svgr/webpack": "^5.4.0",
+    "babel-loader": "^8.1.0",
     "css-loader": "^3.5.3",
     "eslint": "^6.8.0",
     "eslint-config-airbnb-base": "^14.0.0",
