+++ public/stylesheets/dashboard/main.css
@@ -514,96 +514,4 @@ input#header-image {
   height: 100vh;
   border-right: 1px solid #ccc; }
 
-.button {
-  position: relative;
-  cursor: pointer;
-  outline: 0;
-  color: rgba(0, 0, 0, 0.6);
-  font-family: Roboto, 'Helvetica Neue', Arial, Helvetica, sans-serif;
-  padding: 0.78em 1.5em;
-  font-weight: 600;
-  text-align: center;
-  text-decoration: none;
-  border-radius: 0.28rem;
-  user-select: none;
-  overflow: hidden; }
-  .button:hover {
-    color: rgba(0, 0, 0, 0.8); }
-    .button:hover::after {
-      content: '';
-      position: absolute;
-      top: 0px;
-      left: 0px;
-      width: 100%;
-      height: 100%;
-      background: rgba(0, 0, 0, 0.05); }
-  .button:active {
-    color: rgba(0, 0, 0, 0.8); }
-    .button:active::after {
-      content: '';
-      position: absolute;
-      top: 0px;
-      left: 0px;
-      width: 100%;
-      height: 100%;
-      background: rgba(0, 0, 0, 0.1); }
-  .button--variant-text {
-    background: transparent;
-    border: none; }
-  .button--variant-outlined {
-    background: #fff;
-    border: 1px solid #ccc; }
-  .button--variant-filled {
-    background: #e0e1e2;
-    border: none; }
-  .button--size-small {
-    font-size: 0.9rem; }
-  .button--size-medium {
-    font-size: 1rem; }
-  .button--size-large {
-    font-size: 1.15rem; }
-
-.icon-button-container {
-  display: inline-flex;
-  position: relative; }
-  .icon-button-container:hover .tooltip {
-    display: block; }
-
-.icon-button {
-  display: inline-flex;
-  cursor: pointer;
-  padding: 7px;
-  border-radius: 0.5rem;
-  user-select: none;
-  background: transparent;
-  border: transparent;
-  outline: 0;
-  font-family: Roboto, 'Helvetica Neue', Arial, Helvetica, sans-serif;
-  font-size: 1.5rem;
-  color: rgba(0, 0, 0, 0.6);
-  font-weight: 600;
-  text-align: center;
-  text-decoration: none;
-  line-height: 1em; }
-  .icon-button:hover {
-    background: rgba(0, 0, 0, 0.1); }
-  .icon-button:active {
-    background: rgba(0, 0, 0, 0.2); }
-
-.tooltip {
-  display: none;
-  position: absolute;
-  top: -5px;
-  left: 50%;
-  transform-origin: center;
-  transform: translate(-50%, -100%);
-  padding: 5px 8px;
-  background: rgba(97, 97, 97, 0.9);
-  border-radius: 4px;
-  font-family: Roboto, 'Helvetica Neue', Arial, Helvetica, sans-serif;
-  font-size: 0.6rem;
-  font-weight: 500;
-  color: #fff;
-  white-space: nowrap; }
-
 /*# sourceMappingURL=main.css.map */
