+++ sass/dashboard/main.scss
@@ -4,4 +4,3 @@
 @import 'partials/index-section';
 @import 'partials/content-section';
 @import 'partials/gutter';
-@import 'partials/button';
