--- sass/dashboard/partials/_button.scss
@@ -1,124 +1,0 @@
-.button {
-  position: relative;
-  cursor: pointer;
-  outline: 0;
-  // min-height: 1rem;
-  color: rgba(0, 0, 0, 0.6);
-  font-family: Roboto, 'Helvetica Neue', Arial, Helvetica, sans-serif;
-  // vertical-align: baseline;
-  // margin: 0 0.25em 0 0;
-  padding: 0.78em 1.5em;
-  // text-transform: none;
-  // text-shadow: none;
-  font-weight: 600;
-  // line-height: 1em;
-  // font-style: normal;
-  text-align: center;
-  text-decoration: none;
-  border-radius: 0.28rem;
-  // box-shadow: 0 0 0 1px transparent inset, 0 0 0 0 rgba(34,36,38,.15) inset;
-  user-select: none;
-  overflow: hidden;
-
-  &:hover {
-    color: rgba(0, 0, 0, 0.8);
-
-    &::after {
-      content: '';
-      position: absolute;
-      top: 0px; left: 0px;
-      width: 100%; height: 100%;
-      background: rgba(0, 0, 0, 0.05);
-    }
-  }
-
-  &:active {
-    color: rgba(0, 0, 0, 0.8);
-
-    &::after {
-      content: '';
-      position: absolute;
-      top: 0px; left: 0px;
-      width: 100%; height: 100%;
-      background: rgba(0, 0, 0, 0.1);
-    }
-  }
-  
-  &--variant-text {
-    background: transparent;
-    border: none;
-  }
-
-  &--variant-outlined {
-    background: #fff;
-    border: 1px solid #ccc;
-  }
-
-  &--variant-filled {
-    background: #e0e1e2;
-    border: none;
-  }
-
-  &--size-small {
-    font-size: 0.9rem;
-  }
-
-  &--size-medium {
-    font-size: 1rem;
-  }
-
-  &--size-large {
-    font-size: 1.15rem;
-  }
-}
-
-.icon-button-container {
-  display: inline-flex;
-  position: relative;
-
-  &:hover .tooltip {
-    display: block;
-  }
-}
-
-.icon-button {
-  display: inline-flex;
-  cursor: pointer;
-  padding: 7px;
-  border-radius: 0.5rem;
-  user-select: none;
-  background: transparent;
-  border: transparent;
-  outline: 0;
-  font-family: Roboto, 'Helvetica Neue', Arial, Helvetica, sans-serif;
-  font-size: 1.5rem;
-  color: rgba(0, 0, 0, 0.6);
-  font-weight: 600;
-  text-align: center;
-  text-decoration: none;
-  line-height: 1em;
-
-  &:hover {
-    background: rgba(0, 0, 0, 0.1);
-  }
-
-  &:active {
-    background: rgba(0, 0, 0, 0.2);
-  }
-}
-
-.tooltip {
-  display: none;
-  position: absolute;
-  top: -5px; left: 50%;
-  transform-origin: center;
-  transform: translate(-50%, -100%);
-  padding: 5px 8px;
-  background: rgba(97, 97, 97, 0.9);
-  border-radius: 4px;
-  font-family: Roboto, 'Helvetica Neue', Arial, Helvetica, sans-serif;
-  font-size: 0.6rem;
-  font-weight: 500;
-  color: #fff;
-  white-space: nowrap;
-}
