+++ src/dashboard/components/Button/Button.js
@@ -1,5 +1,3 @@
-function _objectWithoutProperties(obj, keys) { var target = {}; for (var i in obj) { if (keys.indexOf(i) >= 0) continue; if (!Object.prototype.hasOwnProperty.call(obj, i)) continue; target[i] = obj[i]; } return target; }
-
 import './Button.scss';
 import React from 'react';
 import PropTypes from 'prop-types';
@@ -7,27 +5,39 @@ import classNames from 'classnames';
 import ButtonBase from '../ButtonBase';
 
 export default function Button(props) {
-  var children = props.children,
-      variant = props.variant,
-      size = props.size,
-      restProps = _objectWithoutProperties(props, ['children', 'variant', 'size']);
+  const {
+    children, variant, size, ...restProps
+  } = props;
 
-  return React.createElement(
-    ButtonBase,
-    Object.assign({
-      className: classNames('button', 'button--variant-' + variant, 'button--size-' + size)
-    }, restProps),
-    children
+  return (
+    <ButtonBase
+      className={classNames(
+        'button',
+        `button--variant-${variant}`,
+        `button--size-${size}`,
+      )}
+      {...restProps}
+    >
+      {children}
+    </ButtonBase>
   );
 }
 
 Button.propTypes = {
   children: PropTypes.string.isRequired,
-  variant: PropTypes.oneOf(['text', 'outlined', 'filled']),
-  size: PropTypes.oneOf(['small', 'medium', 'large'])
+  variant: PropTypes.oneOf([
+    'text',
+    'outlined',
+    'filled',
+  ]),
+  size: PropTypes.oneOf([
+    'small',
+    'medium',
+    'large',
+  ]),
 };
 
 Button.defaultProps = {
   variant: 'text',
-  size: 'medium'
+  size: 'medium',
 };
