+++ src/dashboard/components/IconButton/IconButton.js
@@ -1,5 +1,3 @@
-function _objectWithoutProperties(obj, keys) { var target = {}; for (var i in obj) { if (keys.indexOf(i) >= 0) continue; if (!Object.prototype.hasOwnProperty.call(obj, i)) continue; target[i] = obj[i]; } return target; }
-
 import './IconButton.scss';
 import React from 'react';
 import PropTypes from 'prop-types';
@@ -7,26 +5,15 @@ import ButtonBase from '../ButtonBase';
 import Tooltip from '../Tooltip';
 
 export default function IconButton(props) {
-  var children = props.children,
-      tooltip = props.tooltip,
-      restProps = _objectWithoutProperties(props, ['children', 'tooltip']);
-
-  var iconButton = React.createElement(
-    ButtonBase,
-    Object.assign({ className: 'icon-button' }, restProps),
-    children
-  );
+  const { children, tooltip, ...restProps } = props;
+  const iconButton = <ButtonBase className="icon-button" {...restProps}>{children}</ButtonBase>;
 
   if (!tooltip) return iconButton;
 
-  return React.createElement(
-    Tooltip,
-    { title: tooltip },
-    iconButton
-  );
+  return <Tooltip title={tooltip}>{iconButton}</Tooltip>;
 }
 
 IconButton.propTypes = {
   children: PropTypes.element.isRequired,
-  tooltip: PropTypes.string
+  tooltip: PropTypes.string,
 };
