+++ src/dashboard/components/ProjectListRemoveButtons.js
@@ -1,5 +1,3 @@
-function _toConsumableArray(arr) { if (Array.isArray(arr)) { for (var i = 0, arr2 = Array(arr.length); i < arr.length; i++) { arr2[i] = arr[i]; } return arr2; } else { return Array.from(arr); } }
-
 import React from 'react';
 import ReactDOM from 'react-dom';
 import DeleteIcon from '../icons/delete.svg';
@@ -12,21 +10,20 @@ export default {
   containers: [],
 
   /** Unmount all current containers and mount new react elements again */
-  update: function update() {
-    if (this.containers.length) this.containers.forEach(function (container) {
-      return ReactDOM.unmountComponentAtNode(container);
-    });
+  update() {
+    if (this.containers.length) this.containers.forEach((container) => ReactDOM.unmountComponentAtNode(container));
 
-    var containers = [].concat(_toConsumableArray(document.querySelectorAll('.project-list-remove-button-container')));
+    const containers = [...document.querySelectorAll('.project-list-remove-button-container')];
 
-    containers.forEach(function (container) {
-      ReactDOM.render(React.createElement(
-        IconButton,
-        { tooltip: '\uD504\uB85C\uC81D\uD2B8 \uC0AD\uC81C' },
-        React.createElement(DeleteIcon, null)
-      ), container);
+    containers.forEach((container) => {
+      ReactDOM.render(
+        <IconButton tooltip="프로젝트 삭제">
+          <DeleteIcon />
+        </IconButton>,
+        container,
+      );
     });
 
     this.containers = containers;
-  }
+  },
 };
