--- src/dashboard/components/src/Button/Button.scss
@@ -1,58 +1,0 @@
-.button {
-  position: relative;
-  padding: 0.78em 1.5em;
-  border-radius: 0.28rem;
-  overflow: hidden;
-
-  &:hover {
-    color: rgba(0, 0, 0, 0.8);
-
-    &::after {
-      content: '';
-      position: absolute;
-      top: 0px; left: 0px;
-      width: 100%; height: 100%;
-      background: rgba(0, 0, 0, 0.05);
-    }
-  }
-
-  &:active {
-    color: rgba(0, 0, 0, 0.8);
-
-    &::after {
-      content: '';
-      position: absolute;
-      top: 0px; left: 0px;
-      width: 100%; height: 100%;
-      background: rgba(0, 0, 0, 0.1);
-    }
-  }
-  
-  &--variant-text {
-    background: transparent;
-    border: none;
-  }
-
-  &--variant-outlined {
-    background: #fff;
-    border: 1px solid #ccc;
-  }
-
-  &--variant-filled {
-    background: #e0e1e2;
-    border: none;
-  }
-
-  &--size-small {
-    font-size: 0.9rem;
-  }
-
-  &--size-medium {
-    font-size: 1rem;
-  }
-
-  &--size-large {
-    font-size: 1.15rem;
-  }
-}
-
