--- src/dashboard/components/src/Button/index.js
@@ -1,3 +1,0 @@
-import Button from './Button';
-
-export default Button;
