--- src/dashboard/components/src/ButtonBase/ButtonBase.scss
@@ -1,10 +1,0 @@
-.button-base {
-  font-family: Roboto, 'Helvetica Neue', Arial, Helvetica, sans-serif;
-  font-weight: 600;
-  text-align: center;
-  text-decoration: none;
-  color: rgba(0, 0, 0, 0.6);
-  outline: 0;
-  user-select: none;
-  cursor: pointer;
-}
