--- src/dashboard/components/src/ButtonBase/index.js
@@ -1,3 +1,0 @@
-import ButtonBase from './ButtonBase';
-
-export default ButtonBase;
