--- src/dashboard/components/src/IconButton/index.js
@@ -1,3 +1,0 @@
-import IconButton from './IconButton';
-
-export default IconButton;
