--- src/dashboard/components/src/ProjectFormSubmitButton.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import Button from './Button';
-
-const container = document.querySelector('.project-form-submit-button');
-
-ReactDOM.render(
-  <Button variant="outlined" size="large" type="submit">제출</Button>,
-  container,
-);
