--- src/dashboard/components/src/Tooltip/index.js
@@ -1,3 +1,0 @@
-import Tooltip from './Tooltip';
-
-export default Tooltip;
