+++ webpack.config.js
@@ -13,6 +13,11 @@ module.exports = (env) => ({
   module: {
     rules: [
       {
+        test: /\.js$/,
+        exclude: /node_modules/,
+        loader: 'babel-loader',
+      },
+      {
         test: /\.s[ac]ss$/i,
         use: [
           'style-loader',
